Move progress to utils, update references, add deprecation warning for qt import
