+++ examples/progress_bar_segmentation.py
@@ -6,7 +6,7 @@ import numpy as np
 import napari
 
 from time import sleep
-from napari.qt import progress
+from napari.utils import progress
 from qtpy.QtWidgets import QPushButton, QVBoxLayout, QWidget
 
 from skimage.filters import (
