+++ napari/qt/__init__.py
@@ -7,7 +7,6 @@ from .._qt.qt_viewer import QtViewer
 from .._qt.widgets.qt_tooltip import QtToolTipLabel
 from .._qt.widgets.qt_viewer_buttons import QtStateButton, QtViewerButtons
 from ..utils.translations import trans
-from .progress import progrange, progress
 from .threading import create_worker, thread_worker
 
 
@@ -32,8 +31,6 @@ class QtNDisplayButton(QtStateButton):
 __all__ = (
     'compile_qt_svgs',
     'create_worker',
-    'progress',
-    'progrange',
     'QtStateButton',
     'QtNDisplayButton',
     'QtToolTipLabel',
