+++ napari/utils/__init__.py
@@ -2,6 +2,7 @@ from .colormaps import Colormap
 from .dask_utils import resize_dask_cache
 from .info import citation_text, sys_info
 from .notebook_display import nbscreenshot
+from .progress import progrange, progress
 
 #: dask.cache.Cache, optional : A dask cache for opportunistic caching
 #: use :func:`~.resize_dask_cache` to actually register and resize.
